broken!: Write message length using size_t in the ringbuffer
This doesn't work yet. But I am pretty sure this is what I need to
implement (the Null terminator is not supposed to be used to mark the
end of the reading, instead you should get that information from the
first bytes which represent the message length).

Only thing I still don't quite understand is how to handle messages
overwriting other messages (I am pretty sure that shouldn't even happen,
but it does).